<short summary of the patch>
TODO: Put a short summary on the line above and replace this paragraph
with a longer explanation of this change. Complete the meta-information
with other relevant fields (see below for details). To make it easier, the
information below has been extracted from the changelog. Adjust it or drop
it.

vala-mode-el (0.1-6) unstable; urgency=medium

  * bump policy version
  * bump debhelper version
  * swizzle packaging repo to salsa emacsen-team repo
  * switch priority from extra to optional per policy
  * secure copyright format URL

Gbp-Pq: Name debian-changes